requirements.txt
fastapi
uvicorn
joblib
scikit-learn
numpy
tensorflow-cpu
keras
Pillow